+++ NEWS.md
@@ -1,3 +1,7 @@
+# artpack 0.1.2.9000 (Development Version)
+### **New Functions and Updates**
+* Added `group_slice()` which allows slicing of data frames and tibbles by a group variable.
+
 # artpack 0.1.2 (Development Version)
 ### **New Functions and Updates**
 * Added `group_sample()` which allows sampling of data frames and tibbles by a group variable.